+++ src/resources/views/pages/home.blade.php
@@ -10,11 +10,13 @@
                     <h2><u>Change log</u></h2>
                     <p>Version 0.2 and 0.3 are developed in parallel. The only difference between them is the Laravel version they support. However, this may change in future.</p>
                     <h3>Version 0.3.3.1 and 0.2.3.1</h3>
-                    <p>TinyMCE Improvement for all editors. Image handling improvements.</p>
+                    <p>TinyMCE Improvement for all editors.</p>
                     <p><strong>IMPORTANT:</strong> Refer to UPGRADE.md for the upgrading instructions.</p>
                     <h4>Enhancements:</h4>
                     <ol>
                         <li>Updated TinyMCE to 4.3.8.</li>
+                        <li>TinyMCE: remove paste formatting from MS Office (issue #173)</li>
+                        <li>TinyMCE: ability to insert image from list of uploaded photos (issue #174)</li>
                     </ol>
                     <h4>Bug fixes:</h4>
                     <ol>